requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.1
Flask-Login==0.6.2
Flask-Bootstrap==3.3.7.1
Werkzeug==2.3.7
WTForms==3.0.1
email-validator==2.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.5
gunicorn==20.1.0 